Consolidate & simplify view
