+++ BrowserKit/Sources/ComponentLibrary/Buttons/SecondaryRoundedButtonViewModel.swift
@@ -6,10 +6,10 @@ import Foundation
 
 /// The view model used to configure a `SecondaryRoundedButton`
 public struct SecondaryRoundedButtonViewModel {
-    public let title: String
+    public let title: String?
     public let a11yIdentifier: String
 
-    public init(title: String, a11yIdentifier: String) {
+    public init(title: String?, a11yIdentifier: String) {
         self.title = title
         self.a11yIdentifier = a11yIdentifier
     }
